requirements.txt
requests
pymongo
openai
python-dotenv
langchain
langchain-openai
openai